+++ packages/lib/src/modelShared/prop.ts
@@ -16,7 +16,7 @@ export interface ModelProp<
   TPropCreationValue,
   TTransformedValue,
   TTransformedCreationValue,
-  TIsOptional,
+  TIsRequired,
   TIsId extends boolean = false,
   THasSetter = never,
   TFromSnapshotOverride = never,
@@ -27,7 +27,7 @@ export interface ModelProp<
     $creationValueType: TPropCreationValue
     $transformedValueType: TTransformedValue
     $transformedCreationValueType: TTransformedCreationValue
-    $isOptional: TIsOptional
+    $isRequired: TIsRequired
     $isId: TIsId
     $hasSetter: THasSetter
     $fromSnapshotOverride: TFromSnapshotOverride
@@ -58,7 +58,7 @@ export interface ModelProp<
     TPropCreationValue,
     TTransformedValue,
     TTransformedCreationValue,
-    TIsOptional,
+    TIsRequired,
     TIsId,
     string,
     TFromSnapshotOverride,
@@ -74,7 +74,7 @@ export interface ModelProp<
     TPropCreationValue,
     TTransformedValue,
     TTransformedCreationValue,
-    TIsOptional,
+    TIsRequired,
     TIsId,
     string,
     TFromSnapshotOverride,
@@ -95,7 +95,7 @@ export interface ModelProp<
     TPropCreationValue,
     TTV | Extract<TPropValue, null | undefined>,
     TTV | Extract<TPropCreationValue, null | undefined>,
-    TIsOptional,
+    TIsRequired,
     TIsId,
     THasSetter,
     TFromSnapshotOverride,
@@ -114,7 +114,7 @@ export interface ModelProp<
     TPropCreationValue,
     TTransformedValue,
     TTransformedCreationValue,
-    TIsOptional,
+    TIsRequired,
     TIsId,
     THasSetter,
     FS,
@@ -165,8 +165,8 @@ export interface ModelProps {
   [k: string]: AnyModelProp
 }
 
-export type OptionalModelProps<MP extends ModelProps> = {
-  [K in keyof MP]: MP[K]["_internal"]["$isOptional"] & K
+export type RequiredModelProps<MP extends ModelProps> = {
+  [K in keyof MP]: MP[K]["_internal"]["$isRequired"] & K
 }[keyof MP]
 
 export type ModelPropsToData<MP extends ModelProps> = Flatten<{
@@ -183,12 +183,9 @@ export type ModelPropsToSnapshotData<MP extends ModelProps> = Flatten<{
 // we also don't use Flatten because if we do some generics won't work
 export type ModelPropsToCreationData<MP extends ModelProps> = {
   [k in keyof MP]?: MP[k]["_internal"]["$creationValueType"]
-} & Omit<
-  {
-    [k in keyof MP]: MP[k]["_internal"]["$creationValueType"]
-  },
-  OptionalModelProps<MP>
->
+} & {
+  [k in RequiredModelProps<MP>]: MP[k]["_internal"]["$creationValueType"]
+}
 
 // we don't use O.Optional anymore since it generates unions too heavy
 // also if we use pick over the optional props we will loose the ability
@@ -196,18 +193,15 @@ export type ModelPropsToCreationData<MP extends ModelProps> = {
 export type ModelPropsToSnapshotCreationData<MP extends ModelProps> = Flatten<
   {
     [k in keyof MP]?: ModelPropFromSnapshot<MP[k]> extends infer R ? R : never
-  } & Omit<
-    {
-      [k in keyof MP]: ModelPropFromSnapshot<MP[k]> extends infer R ? R : never
-    },
-    {
+  } & {
+    [k in {
       [K in keyof MP]: IsNeverType<
         MP[K]["_internal"]["$fromSnapshotOverride"],
-        MP[K]["_internal"]["$isOptional"] & K,
-        IsOptionalValue<MP[K]["_internal"]["$fromSnapshotOverride"], K, never>
+        MP[K]["_internal"]["$isRequired"] & K, // no override
+        IsOptionalValue<MP[K]["_internal"]["$fromSnapshotOverride"], never, K> // with override
       >
-    }[keyof MP]
-  >
+    }[keyof MP]]: ModelPropFromSnapshot<MP[k]> extends infer R ? R : never
+  }
 >
 
 export type ModelPropsToTransformedData<MP extends ModelProps> = Flatten<{
@@ -218,14 +212,12 @@ export type ModelPropsToTransformedData<MP extends ModelProps> = Flatten<{
 // also if we use pick over the optional props we will loose the ability
 // to infer generics
 // we also don't use Flatten because if we do some generics won't work
+// we also don't use Omit because if we do some generics won't work
 export type ModelPropsToTransformedCreationData<MP extends ModelProps> = {
   [k in keyof MP]?: MP[k]["_internal"]["$transformedCreationValueType"]
-} & Omit<
-  {
-    [k in keyof MP]: MP[k]["_internal"]["$transformedCreationValueType"]
-  },
-  OptionalModelProps<MP>
->
+} & {
+  [k in RequiredModelProps<MP>]: MP[k]["_internal"]["$transformedCreationValueType"]
+}
 
 export type ModelPropsToSetter<MP extends ModelProps> = Flatten<{
   [k in keyof MP as MP[k]["_internal"]["$hasSetter"] & `set${Capitalize<k & string>}`]: (
@@ -238,7 +230,7 @@ export type ModelIdProp = ModelProp<
   string | undefined,
   string,
   string | undefined,
-  string,
+  never, // not required
   true
 >
 
@@ -270,7 +262,7 @@ export type MaybeOptionalModelProp<TPropValue> = ModelProp<
   TPropValue,
   TPropValue,
   TPropValue,
-  IsOptionalValue<TPropValue, string, never>
+  IsOptionalValue<TPropValue, never, string> // calculate if required
 >
 
 /**
@@ -281,7 +273,7 @@ export type OptionalModelProp<TPropValue> = ModelProp<
   TPropValue | null | undefined,
   TPropValue,
   TPropValue | null | undefined,
-  string
+  never // not required
 >
 
 /**
@@ -348,7 +340,7 @@ export function prop(def?: any): AnyModelProp {
       $creationValueType: null as any,
       $transformedValueType: null as any,
       $transformedCreationValueType: null as any,
-      $isOptional: null as any,
+      $isRequired: null as never,
       $isId: null as never,
       $hasSetter: null as never,
       $fromSnapshotOverride: null as never,